Module drone_riscv::thr

The Threads module.

NOTE This module documentation should be viewed as a continuation of the drone_core documentation.

Re-exports

pub use drone_core::thr::*;

Modules

prelude
The Threads prelude.

Macros

clint
Defines a thread pool driven by CLINT (Core Local Interrupter).

Traits

ThrsInitToken
Threads initialization token.

Functions

init
Initializes the thread system and returns a set of thread tokens.